+++ js/main.js
@@ -1,150 +1,8 @@
-// author
-
-const OFFERS_QUANTITY = 10;
-
-const AVATARS = {
-  min: 1,
-  max: 10,
-};
-
-// offer
-
-const TITLES = [
-  'Tokyo - остановитесь в самом центре',
-  'Асакуса',
-  'Акихабара',
-  'Нихонбаси',
-  'Синджюку',
-  'Акасака',
-  'Гиндза',
-];
-
-const PRICES = {
-  min: 10000,
-  max: 50000,
-};
-
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-
-const ROOMS = {
-  min: 1,
-  max: 3,
-};
-
-const GUESTS = {
-  min: 1,
-  max: 6,
-};
-
-const CHECKIN = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const CHECKOUT = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const DESCRIPTION = [
-  'Очень доступно, чисто, уютно',
-  'Комфортные комнаты',
-  'Весело и ярко',
-  'Отличное расположение',
-  'Номер с основными удобствами и кроватью размера',
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-const LOCATION = {
-  lat: {
-    min: 35.65000,
-    max: 35.70000,
-  },
-  lng: {
-    min: 139.70000,
-    max: 139.80000,
-  },
-  number: 5,
-};
-
-function getRandomInt(min, max) {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.floor(Math.random() * (upper - lower + 1)) + lower;
-
-  return result;
-}
-
-getRandomInt(1,5);
-// https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random  ссылку на источник!
-
-
-function getRandomFloat(min, max, digits = 2) {
-  if (min >= max || min < 0) {
-    throw new Error;
-  }
-  const randomFloat = Math.random() * (max - min) + min;
-  const result = randomFloat.toFixed(digits);
-
-  return result;
-}
-
-getRandomFloat(1, 4, 2);
-
-// Creating New Generator
-
-const createRandomInteger = (min, max) => {
-  const previousValue = [];
-
-  return () => {
-    let  currentValue = getRandomInt (min, max);
-    if (previousValue.length >= (max - min + 1)) {
-      throw new Error;
-    }
-    while (previousValue.includes(currentValue)) {
-      currentValue = getRandomInt (min, max);
-    }
-
-    previousValue.push(currentValue);
-    return currentValue;
-  };
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
-
-const shuffleArray = (array) => {
-  for (let index = array.length - 1; index > 0; index--) {
-    const ji = Math.floor(Math.random() * (index + 1));
-    [array[index], array[ji]] = [array[ji], array[index]];
-  }
-  return array;
-};
-
-const getNewRandomInt = createRandomInteger(AVATARS.min, AVATARS.max);
+import {OFFERS_QUANTITY, AVATARS, TITLES, PRICES, TYPES, ROOMS, GUESTS, CHECKIN, CHECKOUT, FEATURES, DESCRIPTION, PHOTOS, LOCATION} from './data.js';
+import {getRandomInt, getRandomFloat, createRandomInteger, getRandomArrayElement, shuffleArray} from './utils.js';
 
 //  Create new avatar
+const getNewRandomInt = createRandomInteger(AVATARS.min, AVATARS.max);
 
 const createAvatar = () => {
   const newAvatar = getNewRandomInt();
@@ -191,5 +49,3 @@ const newOffer = () => {
 
 const getOffers = (amount) => Array.from({ length: amount }, newOffer);
 getOffers(OFFERS_QUANTITY);
-
-
