+++ package.json
@@ -13,7 +13,7 @@
   "devDependencies": {
     "browser-sync": "2.26.14",
     "eslint": "7.18.0",
-    "eslint-config-htmlacademy": "4.0.0"
+    "eslint-config-htmlacademy": "4.1.0"
   },
   "license": "MIT",
   "dependencies": {},
